+++ core/workflows/circleci.go
@@ -17,6 +17,7 @@ jobs:
       - image: ghcr.io/reliablyhq/cli/cli:latest
         environment:
           RELIABLY_TOKEN: $RELIABLY_TOKEN
+          RELIABLY_ORG: $RELIABLY_ORG
     working_directory: /home
     steps:
       - checkout # check out the code in the project directory