+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Invalid `/relatedto` path for graph visualisation: [#382](https://github.com/reliablyhq/cli/issues/382)
+
 ### Changed
 
 - Added RELIABLY_ORG usage to workflows templates
+
 ## [v0.20.0] - 2021-07-02
 
 ### Added